CheckAuthenticationRegister
