+++ src/test/java/AlloTest/TestClases/CheckingProductCatalog.java
@@ -1,11 +1,8 @@
 package AlloTest.TestClases;
 
 import AlloTest.Pages.CheckingProductCatalogLocators;
-import AlloTest.Pages.PageHomeAllo;
 import ClasesToAllUs.TestInit;
 import org.testng.annotations.Test;
-
-import java.util.List;
 
 public class CheckingProductCatalog extends TestInit {
