+++ src/test/java/AlloTest/TestClases/TestHomeAllo.java
@@ -10,24 +10,6 @@ public class TestHomeAllo extends TestInit {
     PageHomeAllo pageHomeAllo;
 
     @Test
-    public void checkAuthenticationRegister(){
-        pageHomeAllo = new PageHomeAllo(driver);
-        openUrl("https://allo.ua/ru/");
-        pageHomeAllo.singIn().click();
-        pageHomeAllo.loginTab().click();
-        PersonSingIn person = new PersonSingIn ();
-        person.setName("Sergey");
-        person.setTell ("0682223344");
-        person.setEmail ("sergey1234@gmail.com");
-        person.setPassword ("qwerty");
-        pageHomeAllo.enterName().sendKeys (person.getName());
-        pageHomeAllo.enterNumber().sendKeys(person.getTell());
-        pageHomeAllo.enterEmail().sendKeys(person.getEmail());
-        pageHomeAllo.enterPassword().sendKeys(person.getPassword());
-        pageHomeAllo.showPass().click();
-    }
-
-    @Test
     public void checkCategories(){
         pageHomeAllo = new PageHomeAllo(driver);
         openUrl("https://allo.ua/ru/");
